refactor out common stuff to flop_counter and memory_counter
